+++ src/book/read.c
@@ -1,0 +1,3 @@
+fn Term book_read(u64 loc) {
+  return BOOK_HEAP[loc];
+}