+++ src/book/set.c
@@ -1,0 +1,3 @@
+fn void book_set(u64 loc, Term t) {
+  BOOK_HEAP[loc] = t;
+}